Future

+++ server/books/file_service_connector.py
@@ -22,8 +22,11 @@ class FileServiceConnector(CeleryConnector):
 
     def download_file(self, cid: str, key: str, file_type: str):
         filename = f'{uuid.uuid4().hex}.{file_type}.bin'
-        file_path = os.path.join(settings.TEMPORARY_ROOT, filename)
+        if not os.path.exists(settings.BOOK_ROOT):
+            os.mkdir(settings.BOOK_ROOT)
+        file_path = os.path.join(settings.BOOK_ROOT, filename)
         res = self.send_task(FileServiceConfig.TASK_DOWNLOAD_FILE, (file_path, cid, key))
         # remove encrypted file
-        os.remove(file_path)
+        if os.path.exists(file_path):
+            os.remove(file_path)
         return res
